Enh/Fix: Visualization per-thread, better merging of the stacks
With a "per-thread" visualization the stacks are usually separated
from each other by some empty space because of the sampling.

Example:

				Thread-0

    D         D     D       D      E
    C         C     C       C      C    C                          I
    B   B     B     B       B      B    B                     G    H
    A   A     A     A       A      A    A    F    F      F    F    F
|||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||

				Thread-1

      X                                       S
      W   W    V              V               T
      Y   Y    Y	      Y               U        U       U      R
      Z   Z    Z              Z        Z      Z        Z       Z      Z
|||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||

				Thread-2

...

will now appear like that:

				Thread-0

    DDDD      DDDDDDDDDDDDDDDDDDDDDEEEEE
    CCCC      CCCCCCCCCCCCCCCCCCCCCCCCCCCCCCC                      III
    BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB                 GGGGGHHH
    AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAFFFFFFFFFFFFFFFFFFFFFFFFF
|||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||

				Thread-1

      XXXX                                    SSSSSSSSS
      WWWWWWWWWVVVVVVVVVVVVVVVVVVVVVVVV       TTTTTTTTT
      YYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYY       UUUUUUUUUUUUUUUUUUUUUUUUR
      ZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ
|||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||

				Thread-2

...

This breaks the fact that each column is as large as the number of time
it was actually sampled. The width of a column should be interpreted as
the number of time a calling stack was sampled OR the time between two
consecutive samples that saw the given thread working.

Note that with this, the last sample of each thread will not be expanded
to reach the far right of the graph, it will be as large as the number
of times it was last sampled. This is done so that you do not get
confused and are able to tell whether a thread has ended or not.